+++ Kucoin.Net/Clients/SpotApi/KucoinRestClientSpotApiProAccount.cs
@@ -3,6 +3,7 @@ using CryptoExchange.Net.CommonObjects;
 using CryptoExchange.Net.Objects;
 using Kucoin.Net.Converters;
 using Kucoin.Net.Enums;
+using Kucoin.Net.ExtensionMethods;
 using Kucoin.Net.Interfaces.Clients.SpotApi;
 using Kucoin.Net.Objects.Models.Spot;
 using Newtonsoft.Json;
