+++ Kucoin.Net/ExtensionMethods/ServiceCollectionExtensions.cs
@@ -1,21 +1,20 @@
 ﻿using CryptoExchange.Net.Clients;
+using CryptoExchange.Net.Interfaces;
 using Kucoin.Net.Clients;
 using Kucoin.Net.Interfaces;
 using Kucoin.Net.Interfaces.Clients;
 using Kucoin.Net.Objects.Options;
 using Kucoin.Net.SymbolOrderBooks;
-using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Net;
 using System.Net.Http;
-using System.Text.RegularExpressions;
 
-namespace Kucoin.Net
+namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
-    /// Helpers for Kucoin
+    /// Extensions for DI
     /// </summary>
-    public static class KucoinHelpers
+    public static class ServiceCollectionExtensions
     {
         /// <summary>
         /// Add the IKucoinClient and IKucoinSocketClient to the sevice collection so they can be injected
@@ -26,7 +25,7 @@ namespace Kucoin.Net
         /// <param name="socketClientLifeTime">The lifetime of the IKucoinSocketClient for the service collection. Defaults to Singleton.</param>
         /// <returns></returns>
         public static IServiceCollection AddKucoin(
-            this IServiceCollection services, 
+            this IServiceCollection services,
             Action<KucoinRestOptions>? defaultRestOptionsDelegate = null,
             Action<KucoinSocketOptions>? defaultSocketOptionsDelegate = null,
             ServiceLifetime? socketClientLifeTime = null)
@@ -68,19 +67,6 @@ namespace Kucoin.Net
             else
                 services.Add(new ServiceDescriptor(typeof(IKucoinSocketClient), typeof(KucoinSocketClient), socketClientLifeTime.Value));
             return services;
-        }
-
-        /// <summary>
-        /// Validate the string is a valid Kucoin symbol.
-        /// </summary>
-        /// <param name="symbolString">string to validate</param>
-        public static void ValidateKucoinSymbol(this string symbolString)
-        {
-            if (string.IsNullOrEmpty(symbolString))
-                throw new ArgumentException("Symbol is not provided");
-
-            if (!Regex.IsMatch(symbolString, "^((([A-Z]|[0-9]){1,})[-](([A-Z]|[0-9]){1,}))$"))
-                throw new ArgumentException($"{symbolString} is not a valid Kucoin symbol. Should be [BaseAsset]-[QuoteAsset] e.g. ETH-BTC");
         }
     }
 }
