+++ Kucoin.Net/Kucoin.Net.xml
@@ -1872,6 +1872,17 @@
             Failed
             </summary>
         </member>
+        <member name="T:Kucoin.Net.ExtensionMethods.KucoinExtensionMethods">
+            <summary>
+            Extension methods specific to using the Kucoin API
+            </summary>
+        </member>
+        <member name="M:Kucoin.Net.ExtensionMethods.KucoinExtensionMethods.ValidateKucoinSymbol(System.String)">
+            <summary>
+            Validate the string is a valid Kucoin symbol.
+            </summary>
+            <param name="symbolString">string to validate</param>
+        </member>
         <member name="T:Kucoin.Net.Interfaces.Clients.FuturesApi.IKucoinRestClientFuturesApi">
             <summary>
             Client for accessing the Kucoin Futures API. 
@@ -3924,27 +3935,6 @@
             <param name="spotAddress"></param>
             <param name="futuresAddress"></param>
             <returns></returns>
-        </member>
-        <member name="T:Kucoin.Net.KucoinHelpers">
-            <summary>
-            Helpers for Kucoin
-            </summary>
-        </member>
-        <member name="M:Kucoin.Net.KucoinHelpers.AddKucoin(Microsoft.Extensions.DependencyInjection.IServiceCollection,System.Action{Kucoin.Net.Objects.Options.KucoinRestOptions},System.Action{Kucoin.Net.Objects.Options.KucoinSocketOptions},System.Nullable{Microsoft.Extensions.DependencyInjection.ServiceLifetime})">
-            <summary>
-            Add the IKucoinClient and IKucoinSocketClient to the sevice collection so they can be injected
-            </summary>
-            <param name="services">The service collection</param>
-            <param name="defaultRestOptionsDelegate">Set default options for the rest client</param>
-            <param name="defaultSocketOptionsDelegate">Set default options for the socket client</param>
-            <param name="socketClientLifeTime">The lifetime of the IKucoinSocketClient for the service collection. Defaults to Singleton.</param>
-            <returns></returns>
-        </member>
-        <member name="M:Kucoin.Net.KucoinHelpers.ValidateKucoinSymbol(System.String)">
-            <summary>
-            Validate the string is a valid Kucoin symbol.
-            </summary>
-            <param name="symbolString">string to validate</param>
         </member>
         <member name="T:Kucoin.Net.Objects.KucoinApiAddresses">
             <summary>
@@ -9406,18 +9396,33 @@
             Extensions for the ICryptoRestClient and ICryptoSocketClient interfaces
             </summary>
         </member>
-        <member name="M:CryptoExchange.Net.Clients.CryptoClientExtensions.Kucoin(CryptoExchange.Net.Clients.ICryptoRestClient)">
+        <member name="M:CryptoExchange.Net.Clients.CryptoClientExtensions.Kucoin(CryptoExchange.Net.Interfaces.ICryptoRestClient)">
             <summary>
             Get the Kucoin REST Api client
             </summary>
             <param name="baseClient"></param>
             <returns></returns>
         </member>
-        <member name="M:CryptoExchange.Net.Clients.CryptoClientExtensions.Kucoin(CryptoExchange.Net.Clients.ICryptoSocketClient)">
+        <member name="M:CryptoExchange.Net.Clients.CryptoClientExtensions.Kucoin(CryptoExchange.Net.Interfaces.ICryptoSocketClient)">
             <summary>
             Get the Kucoin Websocket Api client
             </summary>
             <param name="baseClient"></param>
+            <returns></returns>
+        </member>
+        <member name="T:Microsoft.Extensions.DependencyInjection.ServiceCollectionExtensions">
+            <summary>
+            Extensions for DI
+            </summary>
+        </member>
+        <member name="M:Microsoft.Extensions.DependencyInjection.ServiceCollectionExtensions.AddKucoin(Microsoft.Extensions.DependencyInjection.IServiceCollection,System.Action{Kucoin.Net.Objects.Options.KucoinRestOptions},System.Action{Kucoin.Net.Objects.Options.KucoinSocketOptions},System.Nullable{Microsoft.Extensions.DependencyInjection.ServiceLifetime})">
+            <summary>
+            Add the IKucoinClient and IKucoinSocketClient to the sevice collection so they can be injected
+            </summary>
+            <param name="services">The service collection</param>
+            <param name="defaultRestOptionsDelegate">Set default options for the rest client</param>
+            <param name="defaultSocketOptionsDelegate">Set default options for the socket client</param>
+            <param name="socketClientLifeTime">The lifetime of the IKucoinSocketClient for the service collection. Defaults to Singleton.</param>
             <returns></returns>
         </member>
     </members>
